server/pyproject.toml
[tool.poetry]
name = "grow"
version = "0.1.0"
description = "Greenhouse Automation Software"
authors = ["Okkema Labs <correos@okkema.org>"]
license = "MIT"
repository = "https://github.com/okkema/grow"

[tool.poetry.dependencies]
python = "^3.10"
aiomysql = "^0.1.1"
databases = "^0.6.0"
fastapi = "^0.79.0"
redis = "^4.3.4"
sse-starlette = "^0.10.3"
python-dotenv = "^0.20.0"
uvicorn = "^0.18.2"
paho-mqtt = "^1.6.1"
asyncio-mqtt = "^0.12.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"
alembic = "^1.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
